LPS-77183 organizations-service - moves OrganizationType to the configuration package
